# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Controle Financeiro - Gabriel e Fernanda
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Income {
  id: number;
  /** gabriel | fernanda */
  person: string;
  /** salario | bonus | plr | decimo_terceiro | outro */
  type: string;
  description: string;
  amount: number;
  month: number;
  year: number;
  createdAt: string;
}

export interface IncomeInput {
  person: string;
  type: string;
  description: string;
  amount: number;
  month: number;
  year: number;
}

export interface IncomeUpdate {
  person?: string;
  type?: string;
  description?: string;
  amount?: number;
  month?: number;
  year?: number;
}

export interface Expense {
  id: number;
  /**
     * gabriel | fernanda | ambos
     * @nullable
     */
  person?: string | null;
  description: string;
  /** Valor total do gasto */
  amount: number;
  /** Custo mensal (para parcelados = amount / totalInstallments) */
  monthlyAmount?: number;
  /** obra | alimentacao | transporte | saude | educacao | lazer | cartao_credito | outros */
  category: string;
  /** fixo | variavel */
  expenseType: string;
  /**
     * credito | debito | pix | dinheiro
     * @nullable
     */
  paymentMethod?: string | null;
  isInstallment?: boolean;
  /** @nullable */
  totalInstallments?: number | null;
  /** @nullable */
  currentInstallment?: number | null;
  startMonth?: number;
  startYear?: number;
  month: number;
  year: number;
  createdAt: string;
}

export interface ExpenseInput {
  person?: string;
  description: string;
  amount: number;
  category: string;
  expenseType: string;
  paymentMethod?: string;
  isInstallment?: boolean;
  /** @nullable */
  totalInstallments?: number | null;
  /** @nullable */
  currentInstallment?: number | null;
  startMonth?: number;
  startYear?: number;
  month: number;
  year: number;
}

export interface ExpenseUpdate {
  person?: string;
  description?: string;
  amount?: number;
  category?: string;
  expenseType?: string;
  paymentMethod?: string;
  isInstallment?: boolean;
  /** @nullable */
  totalInstallments?: number | null;
  /** @nullable */
  currentInstallment?: number | null;
  startMonth?: number;
  startYear?: number;
  month?: number;
  year?: number;
}

export interface CategorySummary {
  category: string;
  total: number;
}

export interface MonthlySummary {
  month: number;
  year: number;
  totalIncome: number;
  totalExpenses: number;
  balance: number;
  fixedTotal: number;
  variableTotal: number;
  obraTotal: number;
  expensesByCategory: CategorySummary[];
  incomeByPerson: CategorySummary[];
}

export interface ObraExpenseItem {
  id: number;
  description: string;
  amount: number;
  monthlyAmount: number;
  isInstallment: boolean;
  /** @nullable */
  totalInstallments?: number | null;
  /** @nullable */
  currentInstallment?: number | null;
  startMonth: number;
  startYear: number;
  /** @nullable */
  paymentMethod?: string | null;
}

export interface ObraSummary {
  totalInvested: number;
  monthlyCommitment: number;
  items: ObraExpenseItem[];
}

export type ListIncomesParams = {
month?: number;
year?: number;
person?: string;
};

export type ListExpensesParams = {
month?: number;
year?: number;
person?: string;
category?: string;
type?: string;
};

export type GetMonthlySummaryParams = {
month: number;
year: number;
};