# lib/api-zod/src/generated/types/categorySummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Controle Financeiro - Gabriel e Fernanda
 * OpenAPI spec version: 0.1.0
 */

export interface CategorySummary {
  category: string;
  total: number;
}