# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Controle Financeiro - Gabriel e Fernanda
 * OpenAPI spec version: 0.1.0
 */

export interface Expense {
  id: number;
  /**
     * gabriel | fernanda | ambos
     * @nullable
     */
  person?: string | null;
  description: string;
  /** Valor total do gasto */
  amount: number;
  /** Custo mensal (para parcelados = amount / totalInstallments) */
  monthlyAmount?: number;
  /** obra | alimentacao | transporte | saude | educacao | lazer | cartao_credito | outros */
  category: string;
  /** fixo | variavel */
  expenseType: string;
  /**
     * credito | debito | pix | dinheiro
     * @nullable
     */
  paymentMethod?: string | null;
  isInstallment?: boolean;
  /** @nullable */
  totalInstallments?: number | null;
  /** @nullable */
  currentInstallment?: number | null;
  startMonth?: number;
  startYear?: number;
  month: number;
  year: number;
  createdAt: string;
}