# lib/api-zod/src/generated/types/expenseInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Controle Financeiro - Gabriel e Fernanda
 * OpenAPI spec version: 0.1.0
 */

export interface ExpenseInput {
  person?: string;
  description: string;
  amount: number;
  category: string;
  expenseType: string;
  paymentMethod?: string;
  isInstallment?: boolean;
  /** @nullable */
  totalInstallments?: number | null;
  /** @nullable */
  currentInstallment?: number | null;
  startMonth?: number;
  startYear?: number;
  month: number;
  year: number;
}