# lib/api-zod/src/generated/types/income.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Controle Financeiro - Gabriel e Fernanda
 * OpenAPI spec version: 0.1.0
 */

export interface Income {
  id: number;
  /** gabriel | fernanda */
  person: string;
  /** salario | bonus | plr | decimo_terceiro | outro */
  type: string;
  description: string;
  amount: number;
  month: number;
  year: number;
  createdAt: string;
}