# lib/api-zod/src/generated/types/incomeUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Controle Financeiro - Gabriel e Fernanda
 * OpenAPI spec version: 0.1.0
 */

export interface IncomeUpdate {
  person?: string;
  type?: string;
  description?: string;
  amount?: number;
  month?: number;
  year?: number;
}