# lib/api-zod/src/generated/types/monthlySummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Controle Financeiro - Gabriel e Fernanda
 * OpenAPI spec version: 0.1.0
 */
import type { CategorySummary } from './categorySummary';

export interface MonthlySummary {
  month: number;
  year: number;
  totalIncome: number;
  totalExpenses: number;
  balance: number;
  fixedTotal: number;
  variableTotal: number;
  obraTotal: number;
  expensesByCategory: CategorySummary[];
  incomeByPerson: CategorySummary[];
}