# lib/api-zod/src/generated/types/obraExpenseItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Controle Financeiro - Gabriel e Fernanda
 * OpenAPI spec version: 0.1.0
 */

export interface ObraExpenseItem {
  id: number;
  description: string;
  amount: number;
  monthlyAmount: number;
  isInstallment: boolean;
  /** @nullable */
  totalInstallments?: number | null;
  /** @nullable */
  currentInstallment?: number | null;
  startMonth: number;
  startYear: number;
  /** @nullable */
  paymentMethod?: string | null;
}