# lib/api-zod/src/generated/types/obraSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Controle Financeiro - Gabriel e Fernanda
 * OpenAPI spec version: 0.1.0
 */
import type { ObraExpenseItem } from './obraExpenseItem';

export interface ObraSummary {
  totalInvested: number;
  monthlyCommitment: number;
  items: ObraExpenseItem[];
}